refactor(controller): 将 getCommentsByPostId 参数类型改为 Integer- 修改了 CommentController 类中 getCommentsByPostId 方法的参数类型 - 从 Long 改为 Integer，以适应前端传入的参数类型 - 优化了参数类型，提高了代码的可用性和维护性

+++ backend/src/main/java/cn/hmck/controller/CommentController.java
@@ -34,7 +34,7 @@ public class CommentController {
      * @return 评论列表
      */
     @GetMapping("/getCommentByPostId")
-    public Result<List<Comment>> getCommentsByPostId(@RequestParam("id") Long postId) {
+    public Result<List<Comment>> getCommentsByPostId(@RequestParam("id") Integer postId) {
         System.out.println("===================================");
         System.out.println("获取评论列表：" + postId);
         System.out.println("===================================");
